Moved a declaration to correct undefined behaviour
See discussion in pull request #7. The conclusion is repeated hereafter.

ISO/IEC 9899:TC3 is perfectly clear about it in 6.2.4 Storage durations
of objects:

The initial value of the object [with automatic storage duration] is
indeterminate. If an initialization is specified for the object, it is
performed each time the declaration is reached in the execution of the
block; otherwise, the value becomes indeterminate each time the
declaration is reached.